Move class auto_buffer to separate file
