+++ xbmc/filesystem/File.cpp
@@ -69,65 +69,6 @@ CFile::~CFile()
 
 //*********************************************************************************************
 
-auto_buffer::auto_buffer(size_t size) : p(NULL), s(0)
-{
-  if (!size)
-    return;
-
-  p = malloc(size);
-  if (!p)
-    throw std::bad_alloc();
-  s = size;
-}
-
-auto_buffer::~auto_buffer()
-{
-  clear();
-}
-
-auto_buffer& auto_buffer::allocate(size_t size)
-{
-  clear();
-  return resize(size);
-}
-
-auto_buffer& auto_buffer::resize(size_t newSize)
-{
-  void* newPtr = realloc(p, newSize);
-  if (!newPtr && newSize)
-    throw std::bad_alloc();
-  p = newPtr;
-  s = newSize;
-  return *this;
-}
-
-auto_buffer& auto_buffer::clear(void)
-{
-  free(p);
-  p = NULL;
-  s = 0;
-  return *this;
-}
-
-auto_buffer& auto_buffer::attach(void* pointer, size_t size)
-{
-  clear();
-  if ((pointer && size) || (!pointer && !size))
-  {
-    p = pointer;
-    s = size;
-  }
-  return *this;
-}
-
-void* auto_buffer::detach(void)
-{
-  void* returnPtr = p;
-  p = NULL;
-  s = 0;
-  return returnPtr;
-}
-
 bool CFile::Copy(const CStdString& strFileName, const CStdString& strDest, XFILE::IFileCallback* pCallback, void* pContext)
 {
   const CURL pathToUrl(strFileName);
